+++ th2_data_services/interfaces/utils/resolver.py
@@ -12,6 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 from abc import ABC, abstractmethod
+from typing import Dict, List, Any, Union
 
 """
 The idea of using resolvers:
@@ -53,109 +54,109 @@ Performance impact:
 class EventFieldResolver(ABC):
     @staticmethod
     @abstractmethod
-    def get_id(event):
+    def get_id(event) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_parent_id(event):
+    def get_parent_id(event) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_status(event):
+    def get_status(event) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_name(event):
+    def get_name(event) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_batch_id(event):
+    def get_batch_id(event) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_is_batched(event):
+    def get_is_batched(event) -> bool:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_type(event):
+    def get_type(event) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_start_timestamp(event):
+    def get_start_timestamp(event) -> Dict[str, int]:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_end_timestamp(event):
+    def get_end_timestamp(event) -> Dict[str, int]:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_attached_messages_ids(event):
+    def get_attached_messages_ids(event) -> List[str]:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_body(event):
+    def get_body(event) -> Any:
         pass
 
 
 class MessageFieldResolver(ABC):
     @staticmethod
     @abstractmethod
-    def get_direction(message):
+    def get_direction(message) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_session_id(message):
+    def get_session_id(message) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_type(message):
+    def get_type(message) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_sequence(message):
+    def get_sequence(message) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_timestamp(message):
+    def get_timestamp(message) -> Dict[str, int]:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_body(message):
+    def get_body(message) -> Union[Dict[str, Any], List[Dict[str, Any]]]:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_body_base64(message):
+    def get_body_base64(message) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_id(message):
+    def get_id(message) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_attached_event_ids(message):
+    def get_attached_event_ids(message) -> List[str]:
         pass
 
     @staticmethod
     @abstractmethod
-    def expand_message(message):
+    def expand_message(message) -> List[Dict[str, Any]]:
         """Extract compounded message into list of individual messages.
 
         Warnings:
@@ -175,27 +176,112 @@ class MessageFieldResolver(ABC):
 class SubMessageFieldResolver(ABC):
     @staticmethod
     @abstractmethod
-    def get_subsequence(message):
+    def get_subsequence(message) -> List[int]:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_type(message):
+    def get_type(message) -> str:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_fields(message):
+    def get_fields(message) -> Dict[str, Any]:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_metadata(message):
+    def get_metadata(message) -> Dict[str, Any]:
         pass
 
     @staticmethod
     @abstractmethod
-    def get_protocol(message):
+    def get_protocol(message) -> str:
+        pass
+
+
+class ExpandedMessageFieldResolver(ABC):
+    @staticmethod
+    @abstractmethod
+    def get_direction(message) -> str:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def get_session_id(message) -> str:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def get_type(message) -> str:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def get_sequence(message) -> str:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def get_timestamp(message) -> Dict[str, int]:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def get_body(message) -> Union[Dict[str, Any], List[Dict[str, Any]]]:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def get_body_base64(message) -> str:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def get_id(message) -> str:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def get_attached_event_ids(message) -> List[str]:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def expand_message(message) -> List[Dict[str, Any]]:
+        """Extract compounded message into list of individual messages.
+
+        Warnings:
+            expand_message function is not backward-compatible.
+            If you use it in your scripts, there is no guarantee that everything will
+            work if you change data-source because different data-sources has different
+            messages structure.
+
+        Args:
+            message: Th2Message
+
+        Returns:
+            Iterable[Th2Message]
+        """
+
+    @staticmethod
+    @abstractmethod
+    def get_subsequence(message) -> List[int]:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def get_fields(message) -> Dict[str, Any]:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def get_metadata(message) -> Dict[str, Any]:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def get_protocol(message) -> str:
         pass